pyproject.toml
[project]
name = "semsb_wati"
version = "0.0.1"
description = "SEMSB WATI WhatsApp Integration for ERPNext"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "pdfplumber",
    "python-dateutil",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.bench.frappe-dependencies]
frappe = ">=15.0.0"
erpnext = ">=15.0.0"